refactor(kubernetes): merge KubernetesV2Configuration into KubernetesConfiguration
