+++ clouddriver-kubernetes/src/main/java/com/netflix/spinnaker/config/KubernetesConfiguration.java
@@ -16,9 +16,18 @@
 package com.netflix.spinnaker.config;
 
 import com.netflix.spectator.api.Registry;
+import com.netflix.spinnaker.cats.module.CatsModule;
+import com.netflix.spinnaker.clouddriver.kubernetes.caching.KubernetesV2Provider;
+import com.netflix.spinnaker.clouddriver.kubernetes.caching.KubernetesV2ProviderSynchronizable;
+import com.netflix.spinnaker.clouddriver.kubernetes.caching.agent.KubernetesV2CachingAgentDispatcher;
 import com.netflix.spinnaker.clouddriver.kubernetes.config.KubernetesConfigurationProperties;
 import com.netflix.spinnaker.clouddriver.kubernetes.health.KubernetesHealthIndicator;
+import com.netflix.spinnaker.clouddriver.kubernetes.model.ManifestProvider;
+import com.netflix.spinnaker.clouddriver.kubernetes.model.NoopManifestProvider;
+import com.netflix.spinnaker.clouddriver.kubernetes.security.KubernetesV2Credentials;
 import com.netflix.spinnaker.clouddriver.security.AccountCredentialsProvider;
+import com.netflix.spinnaker.clouddriver.security.AccountCredentialsRepository;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -45,5 +54,33 @@ public class KubernetesConfiguration {
   public KubernetesHealthIndicator kubernetesHealthIndicator(
       Registry registry, AccountCredentialsProvider accountCredentialsProvider) {
     return new KubernetesHealthIndicator(registry, accountCredentialsProvider);
+  }
+
+  @Bean
+  public KubernetesV2Provider kubernetesV2Provider() {
+    return new KubernetesV2Provider();
+  }
+
+  @Bean
+  public KubernetesV2ProviderSynchronizable kubernetesV2ProviderSynchronizable(
+      KubernetesV2Provider kubernetesV2Provider,
+      AccountCredentialsRepository accountCredentialsRepository,
+      KubernetesV2CachingAgentDispatcher kubernetesV2CachingAgentDispatcher,
+      KubernetesConfigurationProperties kubernetesConfigurationProperties,
+      KubernetesV2Credentials.Factory credentialFactory,
+      CatsModule catsModule) {
+    return new KubernetesV2ProviderSynchronizable(
+        kubernetesV2Provider,
+        accountCredentialsRepository,
+        kubernetesV2CachingAgentDispatcher,
+        kubernetesConfigurationProperties,
+        credentialFactory,
+        catsModule);
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(ManifestProvider.class)
+  public ManifestProvider noopManifestProvider() {
+    return new NoopManifestProvider();
   }
 }
